# docs/swagger.yaml
basePath: /
definitions:
  gin-scalable-api_pkg_response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  internal_modules_role.AssignRoleRequest:
    properties:
      branch_id:
        type: integer
      company_id:
        type: integer
      role_id:
        type: integer
      unit_id:
        type: integer
      user_id:
        type: integer
    required:
    - company_id
    - role_id
    - user_id
    type: object
...
  internal_modules_role.UserRoleAssignmentResponse:
    properties:
      branch_id:
        type: integer
      branch_name:
        type: string
      company_id:
        type: integer
      company_name:
        type: string
      created_at:
        type: string
      id:
        type: integer
      role_id:
        type: integer
      role_name:
        type: string
      unit_id:
        type: integer
      unit_name:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Complete API for ERP with RBAC and API Documentation System. This API provides comprehensive endpoints for managing users, companies, branches, roles, modules, subscriptions, units, applications, and authentication with JWT-based security.

    ## Environments
    - **Development**: http://localhost:8081
    - **Staging**: https://staging-api.huminor.com
    - **Production**: https://api.huminor.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Huminor Console API
  version: "1.0"
paths:
...
  /api/v1/role-management/assign-user-role:
    post:
      consumes:
      - application/json
      description: Menugaskan role ke user dengan scope company, branch, atau unit
      parameters:
      - description: Role assignment data
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/internal_modules_role.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role berhasil ditugaskan
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_role.UserRoleAssignmentResponse'
              type: object
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: User atau role tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Assign role to user
      tags:
      - Role Management
  /api/v1/role-management/bulk-assign-roles:
    post:
      consumes:
      - application/json
      description: Menugaskan role ke multiple users sekaligus
      parameters:
      - description: Bulk role assignment data
        in: body
        name: bulk_assignment
        required: true
        schema:
          $ref: '#/definitions/internal_modules_role.BulkAssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Roles berhasil ditugaskan ke users
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: User atau role tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Bulk assign roles to users
      tags:
      - Role Management
  /api/v1/role-management/role/{roleId}/modules:
    delete:
      consumes:
      - application/json
      description: Menghapus modules dari role
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      - description: Module IDs to remove
        in: body
        name: modules
        required: true
        schema:
          $ref: '#/definitions/internal_modules_role.RemoveRoleModulesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Modules berhasil dihapus
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "400":
          description: Bad request - Invalid role ID atau validation failed
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Role tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Remove modules from role
      tags:
      - Role Management
    post:
      consumes:
      - application/json
      description: Menambahkan modules ke role (append operation)
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      - description: Modules to add
        in: body
        name: modules
        required: true
        schema:
          $ref: '#/definitions/internal_modules_role.AddRoleModulesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Modules berhasil ditambahkan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "400":
          description: Bad request - Invalid role ID atau validation failed
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Role tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Add modules to role
      tags:
      - Role Management
    put:
      consumes:
      - application/json
      description: Mengganti semua permissions/modules dari role (replace operation)
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      - description: Role permissions data
        in: body
        name: permissions
        required: true
        schema:
          $ref: '#/definitions/internal_modules_role.UpdateRolePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permissions berhasil diupdate
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "400":
          description: Bad request - Invalid role ID atau validation failed
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Role tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Update role permissions (replace all)
      tags:
      - Role Management
  /api/v1/role-management/role/{roleId}/users:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar users yang memiliki role tertentu
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      - description: 'Limit jumlah data (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Users berhasil diambil
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "400":
          description: Bad request - Invalid role ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get users by role
      tags:
      - Role Management
  /api/v1/role-management/user/{userId}/access-summary:
    get:
      consumes:
      - application/json
      description: Mendapatkan ringkasan akses user termasuk semua roles dan permissions
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User access summary berhasil diambil
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "400":
          description: Bad request - Invalid user ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get user access summary
      tags:
      - Role Management
  /api/v1/role-management/user/{userId}/role/{roleId}:
    delete:
      consumes:
      - application/json
      description: Menghapus role assignment dari user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      - description: Company ID
        in: query
        name: company_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Role berhasil dihapus dari user
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "400":
          description: Bad request - Invalid ID atau company_id diperlukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: User role assignment tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Remove role from user
      tags:
      - Role Management
  /api/v1/role-management/user/{userId}/roles:
    get:
      consumes:
      - application/json
      description: Mendapatkan semua role assignments dari user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User roles berhasil diambil
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_modules_role.UserRoleAssignmentResponse'
                  type: array
              type: object
        "400":
          description: Bad request - Invalid user ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get user roles
      tags:
      - Role Management
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...